+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/DataApiRequest.java
@@ -6,6 +6,7 @@ import com.yahoo.bard.webservice.data.dimension.Dimension;
 import com.yahoo.bard.webservice.data.dimension.DimensionDictionary;
 import com.yahoo.bard.webservice.data.dimension.DimensionField;
 import com.yahoo.bard.webservice.data.metric.LogicalMetric;
+import com.yahoo.bard.webservice.data.metric.TemplateDruidQuery;
 import com.yahoo.bard.webservice.data.time.Granularity;
 import com.yahoo.bard.webservice.druid.model.builders.DruidFilterBuilder;
 import com.yahoo.bard.webservice.druid.model.filter.Filter;
@@ -21,13 +22,16 @@ import com.yahoo.bard.webservice.web.util.PaginationParameters;
 import org.joda.time.DateTimeZone;
 import org.joda.time.Interval;
 
+import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import javax.ws.rs.core.Response;
 
@@ -414,6 +418,24 @@ public interface DataApiRequest extends ApiRequest {
 
     @Deprecated
     DataApiRequest withFilterBuilder(DruidFilterBuilder filterBuilder);
+
+    /**
+     * The set of referenced dimensions on this ApiRequest.
+     *
+     * @return a set of dimensions
+     */
+    default Set<Dimension> getAllReferencedDimensions() {
+        return Stream.of(
+                getDimensions(),
+                getApiFilters().keySet(),
+                getLogicalMetrics().stream()
+                        .map(LogicalMetric::getTemplateDruidQuery)
+                        .filter(Objects::nonNull)
+                        .map(TemplateDruidQuery::getDimensions)
+                        .flatMap(Collection::stream)
+                        .collect(Collectors.toSet())
+        ).flatMap(Set::stream).collect(Collectors.toCollection(LinkedHashSet::new));
+    }
 
     // CHECKSTYLE:ON
 }