+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/endpoints/SingleDimensionNoRowsFilterDataServletSpec.groovy
@@ -72,7 +72,12 @@ class SingleDimensionNoRowsFilterDataServletSpec extends BaseDataServletComponen
                 "type" : "table"
             },
             "dimensions": [
-                "model"
+                "model",
+                {
+                    "dimension": "misc",
+                    "outputName": "other",
+                    "type": "default"
+                }
             ],
             "filter": {
                 "fields": [